adding ecal candidate filter producers
